UIU-1971-delete-user-with-transactions-check move components to UserDetails, delete unused component
