+++ src/views/UserDetail/components/DeleteUserModal/DeleteUserModal.js
