+++ src/views/UserDetail/components/DeleteUserModal/index.js
@@ -1,0 +1,1 @@
+export { default } from './DeleteUserModal';
