+++ src/views/UserDetail/components/OpenTransactionModal/index.js
@@ -1,0 +1,1 @@
+export { default } from './OpenTransactionModal';
